backend/requirements-render.txt
fastapi==0.110.1
uvicorn==0.25.0
starlette==0.37.2
pydantic==2.12.5
python-dotenv==1.2.2
pandas==3.0.2
openpyxl==3.1.5
requests==2.33.1
